tools/labs: Update tap creation/deletion scripts
Add the `lkt-tap-smbd` tap interface to the `qemu/cleanup-net.sh`
and `qemu/create_net.sh` scirpts. This interface is going to be
used by qemu as a backed to connect to the samba server which
exposes the rootfs.

+++ tools/labs/qemu/cleanup-net.sh
@@ -2,8 +2,11 @@
 
 DNSMASQ=/tmp/dnsmasq
 
-for i in lkt-tap0 lkt-tap1; do
-  if [ $DNSMASQ-$i.pid ]; then
+for i in lkt-tap0 lkt-tap1 lkt-tap-smbd; do
+  if ! ip a s dev &> /dev/null $i; then
+    continue
+  fi
+  if [ -f $DNSMASQ-$i.pid ]; then
     sudo kill `cat $DNSMASQ-$i.pid`
   fi
   sudo rm $DNSMASQ-$i.leases
